Update patch set 2
Patch Set 2:

> Patch Set 1: Verified+1 Code-Review+2
> 
> Ridiculously overdue.  1.5 years for a simple correction, impressive.

From now on, please add Eric and myself to things like this, leave comments/CR, and then wait a period of time to make sure these changes get review... I'm pretty good about being prompt on review.

For example,in this case, I really struggle to think that you covered every variant to get this size? There's likely a variant or reason it was set to that

The _right_ way to do this is to leave it as smallest size, and then use releasetools to resize it on the fly to the max size: https://github.com/LineageOS/android_device_samsung_jf-common/blob/lineage-17.0/releasetools/releasetools.py

I'd almost prefer for safety/compat, we assume that size was set for a reason (some variant we don't know of), and then releasetools it -- it hurts no one and allows for larger GApps installs (likely the goal).

Patch-set: 2
CC: Gerrit User 16185 <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>