Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 1: Verified+1 Code-Review+2
> > > 
> > > Ridiculously overdue.  1.5 years for a simple correction, impressive.
> > 
> > From now on, please add Eric and myself to things like this, leave comments/CR, and then wait a period of time to make sure these changes get review... I'm pretty good about being prompt on review.
> 
> Can I expect the same courtesy from Eric?
> 
> > For example,in this case, I really struggle to think that you covered every variant to get this size? There's likely a variant or reason it was set to that
> 
> Funny, I did check.  One variant has about 10% more allocated for /system, but otherwise the set size is correct size for all other devices as they come from the factory.
> 
> 
> > The _right_ way to do this is to leave it as smallest size, and then use releasetools to resize it on the fly to the max size: https://github.com/LineageOS/android_device_samsung_jf-common/blob/lineage-17.0/releasetools/releasetools.py
> > 
> > I'd almost prefer for safety/compat, we assume that size was set for a reason (some variant we don't know of), and then releasetools it -- it hurts no one and allows for larger GApps installs (likely the goal).
> 
> The justification which had previously been given elsewhere was that there was a plan to break off a chunk to use as /vendor.  Sounds like 17.0 will be using system as / with no tmpfs, in which case there is absolutely no reason to use anything other than the correct size.
> 
> What could be reasonable is some devices system is almost next to userdata, thus there is some potential to steal from system and merge it into userdata.  I've seen no reports of people doing this, but I know the potential exists.

Yes. Eric will do the same, as will I. I'd prefer in the future we get 2 sets of eyes on it. We've got me, yourself, Eric, and Rashed if needed.

If we want to make a vendor partition, which we still can with system mounted as /, we can go that route. If we're going for treble enablement (feasible on 8996), then we do need the vendor partition.

Also, if one variant truely has a bigger partition, we need to add those releasetools I mentioned before we ship anything.

As for what you said about userdata, that's a bad idea IMO, we shouldn't shave system for userdata.

I don't know why you're not in Lineage's slack, would you like an invite? Makes it much easier to talk these things over.

Patch-set: 2